Simulation js

+++ src/commonMain/kotlin/dsl/GroupsIdCreator.kt
@@ -11,8 +11,8 @@ class GroupsIdCreator {
 
     fun addPatternIdCreatorFor(
         group: String,
-        startIndex : Int,
-        formatter: (index: Int) -> String) : PatternIdCreator {
+        startIndex : Long,
+        formatter: (index: Long) -> String) : PatternIdCreator {
         if (group in idCreators) return idCreators[group]!!
         idCreators[group] = PatternIdCreator(startIndex, formatter)
         return idCreators[group]!!
