+++ src/commonMain/kotlin/dsl/ObjectTypeDelegate.kt
@@ -4,7 +4,7 @@ class ObjectTypeDelegate(
     var objectTypeCreator : ObjectTypeCreator
 ) : ObjectTypeAcceptor {
 
-    override fun objectType(label: String, placeNameCreator: ((placeIndexForType: Int) -> String)): ObjectTypeDSL {
+    override fun objectType(label: String, placeNameCreator: ((placeIndexForType: Long) -> String)): ObjectTypeDSL {
         return objectTypeCreator.createObjectType(label, placeNameCreator)
     }
